log tap/long preess on each Dice or tap/long press on power switcher

+++ Milestone Project 16-18 - Dice/Dice/RollingLog.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct RollingLog: Codable, Comparable {
-    let dices: Int
-    let posibilities: Int
+    let dices: Double
+    let posibilities: Double
     let result: String
     let sumOfResult: Int
     let highestResult: Int
